backend/backend/requirements.txt
certifi==2019.11.28
chardet==3.0.4
crypto==1.4.1
Django==2.2.7
django-cors-headers==3.2.0
django-easy-pdf==0.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
html5lib==1.0.1
idna==2.8
Naked==0.1.31
Pillow==6.2.1
pycryptodome==3.9.4
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2019.3
PyYAML==5.3
reportlab==3.5.42
requests==2.22.0
shellescape==3.8.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
webencodings==0.5.1
xhtml2pdf==0.2.4
